refactor(state): split AppState into modules
